Refactoring cpuinfo.

+++ src/CpuInfo/main.cpp
@@ -10,7 +10,7 @@ int main()
     const CpuInfo& cpu = get_cpu_info();
     cout << "Arch         : " << arch_to_string(cpu.m_arch, cpu.m_level) << endl;
     cout << "Brand name   : " << cpu.m_brandString << endl;
-    cout << "F,M,S,xF,xS  : " << hex << cpu.m_family << ", " << cpu.m_model << ", " << cpu.m_stepping << ", " << cpu.m_ext_family << ", " << cpu.m_ext_model << endl;
+    cout << "F,M,S,xF,xS  : " << hex << (int)cpu.m_family << ", " << (int)cpu.m_model << ", " << (int)cpu.m_stepping << ", " << (int)cpu.m_ext_family << ", " << (int)cpu.m_ext_model << endl;
     cout << "Codename     : " << cpu.m_codeName << endl;
     cout << "Class        : " << class_to_string(cpu.m_class) << endl;
     cout << "Hybrid       : " << (cpu.m_hybrid ? "true" : "false") << endl;
